pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-content-optimizer"
description = "A desktop application for optimizing YouTube content using AI"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "flet",
    "ffmpeg-python",
    "opencv-python",
    "faster-whisper",
    "requests",
    "torch",
]
dynamic = ["version"]

[project.scripts]
youtube-optimizer = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/youtube-content-optimizer"
Repository = "https://github.com/yourusername/youtube-content-optimizer"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "dirty-tag"